pyproject.toml
[project]
name = "carla-rl-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "carla==0.9.16",
    "numpy",
    "gymnasium",
    "stable-baselines3[extra]",
    "PyYAML",
    "torch"
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }